Move load_accounts to runtime (#34017)
* Move load_accounts to runtime

---------
